move `SourceStreamChunkBuilder` to separate file
